requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
tushare
finnhub-python
parsel
requests
python-binance
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1 
